Added custom groups
Added custom groups to allow files sharing across local users on external storage (or internal on inverted kernel)

Adding correct UID to GID will be at owner's discretion. Including the GID here is just to bypass the fact "addgroup" doesn't exist on this rom, and if it existed we don't know if it would survive updating system.

This is untested but worst case should be the end user cannot use the groups.